# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type CreateConsultationInputPreferredContact =
  (typeof CreateConsultationInputPreferredContact)[keyof typeof CreateConsultationInputPreferredContact];

export const CreateConsultationInputPreferredContact = {
  phone: "phone",
  whatsapp: "whatsapp",
  email: "email",
} as const;

export interface CreateConsultationInput {
  /**
   * @minLength 2
   * @maxLength 120
   */
  fullName: string;
  /**
   * @minLength 7
   * @maxLength 40
   */
  phone: string;
  /** @maxLength 200 */
  email?: string;
  /** @maxLength 120 */
  city?: string;
  /** @maxLength 80 */
  practiceArea?: string;
  /**
   * @minLength 10
   * @maxLength 4000
   */
  matter: string;
  preferredContact?: CreateConsultationInputPreferredContact;
}

export interface CreateConsultationResponse {
  id: number;
  message: string;
}

export interface Consultation {
  id: number;
  fullName: string;
  phone: string;
  email?: string | null;
  city?: string | null;
  practiceArea?: string | null;
  matter: string;
  preferredContact?: string | null;
  status: string;
  createdAt: string;
}

export interface ConsultationsList {
  items: Consultation[];
}

export interface ErrorResponse {
  message: string;
}