# lib/api-zod/src/generated/types/consultation.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Consultation {
  id: number;
  fullName: string;
  phone: string;
  email?: string | null;
  city?: string | null;
  practiceArea?: string | null;
  matter: string;
  preferredContact?: string | null;
  status: string;
  createdAt: Date;
}