# lib/api-zod/src/generated/types/consultationsList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Consultation } from "./consultation";

export interface ConsultationsList {
  items: Consultation[];
}