# lib/api-zod/src/generated/types/createConsultationInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateConsultationInputPreferredContact } from "./createConsultationInputPreferredContact";

export interface CreateConsultationInput {
  /**
   * @minLength 2
   * @maxLength 120
   */
  fullName: string;
  /**
   * @minLength 7
   * @maxLength 40
   */
  phone: string;
  /** @maxLength 200 */
  email?: string;
  /** @maxLength 120 */
  city?: string;
  /** @maxLength 80 */
  practiceArea?: string;
  /**
   * @minLength 10
   * @maxLength 4000
   */
  matter: string;
  preferredContact?: CreateConsultationInputPreferredContact;
}